docs/mmos/pipeline/requirements.txt
anthropic>=0.18.0
click>=8.1.0
pyyaml>=6.0
